+++ VersionHistory.md
@@ -9,6 +9,7 @@ Prefix the description of the change with `[major]`, `[minor]`, or `[patch]` in 
 * [minor] Add `FindFilesFrom` and `FindDirectoriesFrom` to `BuildUtility`.
 * [minor] Globs should use case-insensitive matching.
 * [patch] Don't return the same found file/directory twice.
+* [minor] `BuildUtility.CopyFiles` uses globs to copy files.
 
 ## Released